librsvg: Autospec creation for version 2.45.5
The biggest changes for building are the configure32 and make32_args
and the switching to the make dist upstream source tar. Due to vendor
content being included we drop the custom created vendor sources.

The 32bit build went through a number of modifications in order to
avoid cross compiling and mostly every change associated with it is a
hack that probably should be addressed better upstream (TODO).